Fix date boundaries
In this example:

    mrodwin — 2/28/25, 21:31
    Daily Octordle #1131
    7️⃣🔟
    5️⃣8️⃣
    🟥🟥 oof
    🟥🕐
    Score: 85

The score shows in the March (3/1/25 - 3/31/25) listing
because 2/28/25, 21:31 (EST) is understood as 3/1/25, 02:31 (UTC).

This problem generalizes if anyone dumps a bunch of scores in bulk into a channel.
So, we need to track the day of a puzzle and filter that way.

Now, I could write a function ("Daily Octordle", "1131") -> "2025-02-28".
My old ByteBinder code did that. However, now that I'm filtering by date,
I either need to put a ("Daily Octordle", "1131", "2025-02-28") tuple somewhere to join against,
or route the SQL execution plan somewhere I can execute the function.